[Refactor] CompareViewViewController와 RouteDataModel 분리, enum MapName, SearchOption 별도 파일에 분리
[Refactor] CompareViewViewController와 RouteDataModel 분리

enum MapName, SearchOption 별도 파일에 분리
